requirements.txt
openai-agents>=0.1.0
openai>=1.12.0
flask>=3.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
gunicorn>=21.2.0
